Bug 1577766 - Move CompositeOps into display_list_flattener.rs. r=gw
Differential Revision: https://phabricator.services.mozilla.com/D44171

[wrupdater] From https://hg.mozilla.org/mozilla-central/rev/cc1936e4a808d8f491334ed6011fe98cd57fe89a
