+++ webrender/src/render_task_graph.rs
@@ -7,8 +7,8 @@
 //! Code associated with creating specific render tasks is in the render_task
 //! module.
 
-use api::{ColorF, BorderStyle, FilterPrimitive, MixBlendMode, PipelineId, PremultipliedColorF};
-use api::{DocumentLayer, FilterData, ImageFormat, LineOrientation};
+use api::{ColorF, BorderStyle, PipelineId, PremultipliedColorF};
+use api::{DocumentLayer, ImageFormat, LineOrientation};
 use api::units::*;
 use crate::batch::{AlphaBatchBuilder, AlphaBatchContainer, BatchTextures, ClipBatcher, resolve_image, BatchBuilder};
 use crate::clip::ClipStore;
@@ -19,7 +19,7 @@ use crate::frame_builder::FrameGlobalResources;
 use crate::gpu_cache::{GpuCache, GpuCacheAddress};
 use crate::gpu_types::{BorderInstance, SvgFilterInstance, BlurDirection, BlurInstance, PrimitiveHeaders, ScalingInstance};
 use crate::gpu_types::{TransformData, TransformPalette, ZBufferIdGenerator};
-use crate::internal_types::{CacheTextureId, FastHashMap, LayerIndex, SavedTargetIndex, Swizzle, TextureSource, Filter};
+use crate::internal_types::{CacheTextureId, FastHashMap, LayerIndex, SavedTargetIndex, Swizzle, TextureSource};
 use crate::picture::{RecordedDirtyRegion, SurfaceInfo};
 use crate::prim_store::gradient::GRADIENT_FP_STOPS;
 use crate::prim_store::{PrimitiveStore, DeferredResolve, PrimitiveScratchBuffer, PrimitiveVisibilityMask};
@@ -1685,39 +1685,6 @@ impl RenderPass {
                 );
             }
         }
-    }
-}
-
-#[derive(Debug, Clone, Default)]
-pub struct CompositeOps {
-    // Requires only a single texture as input (e.g. most filters)
-    pub filters: Vec<Filter>,
-    pub filter_datas: Vec<FilterData>,
-    pub filter_primitives: Vec<FilterPrimitive>,
-
-    // Requires two source textures (e.g. mix-blend-mode)
-    pub mix_blend_mode: Option<MixBlendMode>,
-}
-
-impl CompositeOps {
-    pub fn new(
-        filters: Vec<Filter>,
-        filter_datas: Vec<FilterData>,
-        filter_primitives: Vec<FilterPrimitive>,
-        mix_blend_mode: Option<MixBlendMode>
-    ) -> Self {
-        CompositeOps {
-            filters,
-            filter_datas,
-            filter_primitives,
-            mix_blend_mode,
-        }
-    }
-
-    pub fn is_empty(&self) -> bool {
-        self.filters.is_empty() &&
-            self.filter_primitives.is_empty() &&
-            self.mix_blend_mode.is_none()
     }
 }
